requirements.txt
Django==2.2.7
djangorestframework==3.9.4
django-rest-swagger==2.2.0
django-cors-headers==3.2.1
django-environ==0.4.4
django-log-request-id==1.4.1
drf-yasg==1.17.0
psycopg2==2.7.4
